Add explicit GC links from structures to fields
As reported in <https://github.com/cryptosense/pkcs11/issues/10>, Ctypes
does not expose that fields are reachable through a structure value.

Therefore, it is unsafe to store the result of `allocate` (or
`allocate_n`) in a structure using `setf`.

This adds a `add_gc_link` function that attaches a finalizer to a value,
with a closure that holds a reference to the pointed value until the
finalizer runs.

Closes #10