Wrap `add_gc_link` in a type-safe module
This adds a `Reachable_ptr` module that is a type-safe wrapper around
`add_gc_link`. It guarantees that values of this type were
attached to the parent structure value using `add_gc_link`.

Changing the declarations from `Ctypes.ptr` to `Reachable_ptr.typ`
reveal a couple more instances of this problem: this also fixes them.