Remove "raw payload" modules
It was used only for `CKM_CS_UNKNOWN`. But when the mechanism type is
not known, it is impossible to know the size of the parameters object
(it was assumed to be a string).

Instead this uses only the ulong identifying the mechanism type.

The only downside to this is that make/view is not idempotent (the
params field becomes `NULL`), but that was not something we relied on.