Deprecate the "stubs" load mode
This mode relies on some C code with globals that is linked into the
library. The only advantage of it is that it does not rely on `libffi`,
but that library is available virtually everywhere, even in obscure Unix
systems.

This load mode will be removed in a future release, and in particular
`--indirect` will be removed as well.